move guard to VerificationDocumentStartAction
